refac: Refactorize `create_test` and `print_test_info`. #1864
Now when a user wants to parse a single test, the info is printed or dumped into a file if an output directory is specified, but now reutilizing the `dump_output` method to create the file and the print method only prints the related test info.

The test info that is printed now is the same that is dumped.