feat(#3773): add engine's kvdb api dump and delete tests
The following kvdb's api tests with their cases have been added:
- dump subcommand
- delete subcommand

The documentation has been updated and added to these new tests.

Besides, two tests are skipped because the processes method is not returning anything and the unicode case for dump is also skipped because something is happening with the decoding.