refactor(#3773): Update tests after commands changes
The dev team did some changes to the commands and API, so it was necessary to refactor expected outputs, and other things.
There are many xfails because the list and dump commands are not working as expected right now when listing empty lists and dumping empty kvdbs.
Also, there are some cases skipped because unicode escaping is case-sensitive in Python. We need to fix it somehow.